backend/requirements.txt
fastapi
uvicorn
sqlalchemy
google-generativeai
python-multipart
pdf2image
python-dotenv
aiofiles
